+++ src/os/shared/src/osapi-file.c
@@ -522,7 +522,7 @@ int32 OS_cp(const char *src, const char *dest)
         wr_total = 0;
         while (wr_total < rd_size)
         {
-            wr_size = OS_write(file2, &copyblock[wr_total], rd_size - wr_total);
+            wr_size = OS_write(file2, &copyblock[wr_total], (uint32)(rd_size - wr_total));
             if (wr_size < 0)
             {
                 return_code = wr_size;
